add request body for create thread
